New feature added: TekScan calibration code
A new data aquisition device is added in the analysis code: the TekScan
handle. The sensors of TekScan come uncalibrated, that's why I first need
to aquire some measurements with a known load applied on the sensor and do
a curve fitting on the measurements. This is what this code is doing.

I should also code something that converts the TekScan raw measurements to
measurements with Units, once you apply the calibration matrix of each
sensor.