Corrected calibration of data
A logic error was fixed, in order to do the calibration of the data
correctly (the xdata and the ydata were swapped).

Furthermore, the way the load on each sensel is calculated was changed,
so that the actual pressure (instead of the force) is used for the
calibration.

Last but not least, I implemented a feature that tries to find the global
minimum of the optimisation function during the curve fitting