Semi-Automatic selection of the order of the fitted polynomial
The user selectes from a prompt the order of the polynomial that will be
used in order to fit the data. The function needs to be written manually,
but all the rest are done automatically (in fact, the function can be
derived automatically, it just takes a long time to evaluate it).

Furthermore, the positioning of the figures is harmonised for better and
quicker visualisations